% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{methods}
\name{as.array}
\alias{as.array}
\alias{dim,antsImage-method}
\alias{is.na,antsImage-method}
\alias{.}
\alias{as.numeric,antsImage-method}
\alias{as.matrix,antsImage-method}
\alias{as.matrix.antsImage}
\alias{as.array,antsImage-method}
\alias{as.array.antsImage}
\title{Coerce antsImage objects to array}
\usage{
\S4method{dim}{antsImage}(x)

\S4method{is.na}{antsImage}(x)

\S4method{as.numeric}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.matrix}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\method{as.matrix}{antsImage}(x, ..., mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.array}{antsImage}(x, ..., mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\method{as.array}{antsImage}(x, ..., mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))
}
\arguments{
\item{x}{object of class \code{antsImage} or \code{antsMatrix}}

\item{mask}{a logical vector/array or binary antsImage object}

\item{region}{a \code{antsRegion} object}

\item{...}{additional arguments passed to functions}
}
\description{
Converts antsImage, antsImage object to different data types
}
